automaton-auditor/pyproject.toml
[project]
name = "automaton-auditor"
version = "1.0.0"
description = "Enterprise-grade multi-agent code auditing system"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "docling>=1.0.0",
    "pdf2image>=1.16.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "langsmith>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
